requirements.txt
Django==4.1.2
djangorestframework==3.14.0
drf_yasg==1.21.4
eth_account==0.5.9
Jinja2==3.1.2
pandas==1.4.3
prometheus_client==0.14.1
PyYAML==6.0
requests==2.28.1
tablib==3.2.1
web3==5.30.0
django-celery-results==2.4.0
django-cors-headers==3.13.0
mysqlclient==2.1.1
django-filter==22.1
